caughtscreen
all my pushes relate to enemy script editing, a object that handles pop up screen based on collisions(death), a new script that runs in this object and is called when a collision happens, also will servee as the template for a future options screen when we die. Made a level called LVL1Reload thats where I made my changes and i cloned bhaviks branch and his lighting scene to incorporate my changes.